theme: Add original theming by brother
This bitmaps are contained in the D600 COLOR PD3 file and have been
extracted from there for further processing.

Each file has the table index in its name and the original image
dimensions.

Also node the `header.json` which contains the information transported
in the header of the PD3 for easier combination later on.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>